+++ include/envs.h
@@ -101,6 +101,8 @@ static char *afl_environment_variables[] = {
     "AFL_LLVM_LTO_DONTWRITEID",
     "AFL_NO_ARITH",
     "AFL_NO_BUILTIN",
+    "AFL_NO_COLOR",
+    "AFL_NO_COLOUR",
     "AFL_NO_CPU_RED",
     "AFL_NO_FORKSRV",
     "AFL_NO_UI",