+++ instrumentation/compare-transform-pass.so.cc
@@ -48,6 +48,7 @@
 #include <set>
 #include "afl-llvm-common.h"
 
+unsigned char colored = 1;
 using namespace llvm;
 
 namespace {