+++ instrumentation/split-switches-pass.so.cc
@@ -50,6 +50,7 @@
 #include <set>
 #include "afl-llvm-common.h"
 
+unsigned char colored = 1;
 using namespace llvm;
 
 namespace {