+++ src/afl-as.c
@@ -53,6 +53,7 @@
 #include <sys/wait.h>
 #include <sys/time.h>
 
+u8  colored = 1;
 static u8 **as_params;              /* Parameters passed to the real 'as'   */
 
 static u8 *input_file;              /* Originally specified input file      */