+++ src/afl-common.c
@@ -45,6 +45,7 @@
 u8  be_quiet = 0;
 u8 *doc_path = "";
 u8  last_intr = 0;
+u8  colored = 1;
 
 void detect_file_args(char **argv, u8 *prog_in, u8 *use_stdin) {