+++ test/unittests/unit_maybe_alloc.c
@@ -21,6 +21,7 @@ extern void mock_assert(const int result, const char* const expression,
 #define assert(expression) \
     mock_assert((int)(expression), #expression, __FILE__, __LINE__);
 #include "alloc-inl.h"
+unsigned char colored = 0;
 
 void __wrap_exit(int status);
 /* remap exit -> assert, then use cmocka's mock_assert